+++ applications/portfolio_manager/src/portfolio_manager/rebalance.py
@@ -45,7 +45,7 @@ _MINIMUM_PAIR_PRICE_ROWS = 30
 
 _PRIOR_PORTFOLIO_SCHEMA: dict[str, type] = {
     "ticker": pl.String,
-    "timestamp": pl.Float64,
+    "timestamp": pl.Int64,
     "side": pl.String,
     "dollar_amount": pl.Float64,
     "action": pl.String,
